Synovial differential expression of modules across pathotypes in PEAC cohort

A dataset containing the relative differential expression parameters comparing module expression between different pathotype groups for 81 synovial biopsies from the PEAC cohort.

syn_mod_pvalues

Format

A data frame with 346 rows representing modules and 10 columns for each statistical parameter:

Module

The gene name

Lymphoid-Fibroid p.value

pvalue from lymphoid vs fibroid comparison

Lymphoid-Fibroid q.value

adjusted pvalue from lymphoid vs fibroid comparison

Lymphoid-Fibroid logFC

logarithmic fold change from lymphoid vs fibroid comparison

Lymphoid-Myeloid p.value

pvalue from lymphoid vs myeloid comparison

Lymphoid-Myeloid q.value

adjusted pvalue from lymphoid vs myeloid comparison

Lymphoid-Myeloid logFC

logarithmic fold change from lymphoid vs myeloid comparison

Myeloid-Fibroid p.value

pvalue from myeloid vs fibroid comparison

Myeloid-Fibroid q.value

adjusted pvalue from myeloid vs fibroid comparison

Myeloid-Fibroid logFC

logarithmic fold change from myeloid vs fibroid comparison

Source

https://www.cell.com/cell-reports/fulltext/S2211-1247(19)31007-1